succefully implemented the route to home page
