+++ src/components/LandingPage.js
@@ -1,0 +1,11 @@
+import { Link } from 'react-router-dom';
+
+const LandingPage = (() => {
+    return(
+        <>
+        <Link to='/welcome'>Click here to begin</Link>
+        </>
+    )
+})
+
+export default LandingPage;
